+++ sneakctl_server/flaskr/app.py
@@ -1,6 +1,9 @@
 from flask import Flask
 
+from sneakctl_server.flaskr.routes.file import files_blueprint
+from sneakctl_server.flaskr.routes.fswatch import fswatch_blueprint
 from sneakctl_server.flaskr.routes.process import process_blueprint
+from sneakctl_server.flaskr.routes.tcptracer import tcptracer_blueprint
 from sneakctl_server.flaskr.settings import prod as prod_cfg, dev as dev_cfg
 from sneakctl_server.flaskr import db
 from sneakctl_server.flaskr.routes.execsnoop import execsnoop_blueprint
@@ -28,6 +31,9 @@ def create_app():
     app.register_blueprint(db_blueprint)
     app.register_blueprint(execsnoop_blueprint)
     app.register_blueprint(process_blueprint)
+    app.register_blueprint(fswatch_blueprint)
+    app.register_blueprint(files_blueprint)
+    app.register_blueprint(tcptracer_blueprint)
 
     db.init_app(app)