+++ sneakctl_server/flaskr/routes/fswatch.py
@@ -4,29 +4,29 @@ from werkzeug.exceptions import BadRequest, BadRequestKeyError
 from sneakctl_server.core.utils import str_to_bool
 from sneakctl_server.core.core import interface
 
-execsnoop_blueprint = Blueprint('execsnoop_blueprint', __name__, url_prefix='/execsnoop')
+fswatch_blueprint = Blueprint('fswatch_blueprint', __name__, url_prefix='/fswatch')
 
 
-@execsnoop_blueprint.route('/status', methods=['GET'])
+@fswatch_blueprint.route('/status', methods=['GET'])
 def status():
     """
-    Get status on all execsnoop processes. Use URL parameter 'reload' to reload the instances and get the current
+    Get status on all fswatch processes. Use URL parameter 'reload' to reload the instances and get the current
     status. It defaults to true.
 
     :return: list of process_info dicts
     """
     status_response = {'response': {}}
     if str_to_bool(request.args.get('reload', default='true')):
-        interface.execsnoop_adapter.load_instances()
+        interface.fswatch_adapter.load_instances()
 
-    status_response['request'] = '/execsnoop/status'
-    status_response['response']['process_instances'] = interface.execsnoop_adapter.to_json()
+    status_response['request'] = '/fswatch/status'
+    status_response['response']['process_instances'] = interface.fswatch_adapter.to_json()
     status_response['response']['count'] = len(status_response['response']['process_instances'])
 
     return status_response
 
 
-@execsnoop_blueprint.route('/services', methods=['GET'])
+@fswatch_blueprint.route('/services', methods=['GET'])
 def services():
     """
     Get all execsnoop systemd services, that are manageable by this server.
@@ -34,22 +34,22 @@ def services():
     :return: list of services
     """
     if request.method == 'GET':
-        units = interface.systemd_adapter.get_all_systemd_units('execsnoop@')
+        units = interface.systemd_adapter.get_all_systemd_units('fswatch@')
         return {
-            'request': '/execsnoop/services',
+            'request': '/fswatch/services',
             'response': units
         }
 
 
-@execsnoop_blueprint.route('/services/<name>', methods=['GET', 'POST'])
+@fswatch_blueprint.route('/services/<name>', methods=['GET', 'POST'])
 def services_modify(name):
     """
-    Get, stop or restart an execsnoop service by name or all (<name> == '_all')
+    Get, stop or restart an fswatch service by name or all (<name> == '_all')
 
     json body:
     { "action": on of ["start", "stop", "restart"] }
 
-    :param name: execsnoop service name or '_all' for addressing all execsnoop services
+    :param name: fswatch service name or '_all' for addressing all execsnoop services
     :return: status of changes made or requested service info
     """
     ALL_SERVICES = '_all'
@@ -57,7 +57,7 @@ def services_modify(name):
     DEFAULT_STATUS_CODE = 200
 
     contents = {}
-    response = {'request': f'/execsnoop/services/{name}'}
+    response = {'request': f'/fswatch/services/{name}'}
     msg = {}
     status_code = DEFAULT_STATUS_CODE
 
@@ -83,12 +83,12 @@ def services_modify(name):
         return {}, 400
 
     if status_code == DEFAULT_STATUS_CODE:
-        units = interface.systemd_adapter.get_all_systemd_units('execsnoop@')
+        units = interface.systemd_adapter.get_all_systemd_units('fswatch@')
         if request.method == 'GET':
             # resolve the GET request and return a single service
             if name == ALL_SERVICES:
                 # ALL_SERVICES is allowed only for POST method
-                msg = {'message': "info: use the designated API endpoint '/execsnoop/services' or use POST"}
+                msg = {'message': "info: use the designated API endpoint '/fswatch/services' or use POST"}
             else:
                 found_service = [unit for unit in units if unit['name'] == name]
                 if len(found_service) > 0: