+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 
 ## [Unreleased]
+_
+
+## [0.1.0]
+### Added
+- retrieve all execsnoop services
+- execsnoop/services/<name> for start/stop/restart operations
+    - separate systemd management class
+
+- default route '/' to retrieve version and other future info
+- bug fixes
+- bulk request '/execsnoop/services/_all' to manage all services
+- fswatch and tcptracer API endpoints same as for execsnoop
+- reset-failed systemd service after each operation execution
+- file stats API
+
+### Modified
+- move main object adapters (execsnoop and systemd adapters) to a separate class/file
+- execsnoop service management
+- unified the API responses for execsnoop blueprint
+
+### Removed
+- execsnoop/load api call and integrated it to the execsnoop/status api call
 
 ## [0.0.4]
 ### Modified
@@ -36,9 +58,9 @@
 
 - serialize all existing execsnoop instances
 
-[Unreleased]: https://github.com/tomas321/sneakctl_server/compare/0.0.4...develop
+[Unreleased]: https://github.com/tomas321/sneakctl_server/compare/0.1.0...develop
+[0.1.0]: https://github.com/tomas321/sneakctl_server/compare/0.0.4...0.1.0
 [0.0.4]: https://github.com/tomas321/sneakctl_server/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/tomas321/sneakctl_server/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/tomas321/sneakctl_server/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/tomas321/sneakctl_server/releases/tag/0.0.1
-